requirements.txt
# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 의미 분석 (SBERT)
sentence-transformers>=2.2.0
torch>=2.0.0

# 자연어 처리
keybert>=0.8.0  # 키워드 추출
bertopic>=0.15.0  # 토픽 모델링 및 클러스터링
scikit-learn>=1.3.0  # K-means 클러스터링 (BERTopic에서 사용)
umap-learn>=0.5.0  # 차원 축소 (BERTopic 의존성)
hdbscan>=0.8.0  # 밀도 기반 클러스터링 (BERTopic 의존성)

# 그래프 DB
neo4j>=5.0.0
py2neo>=2021.2.3

# API 요청
requests>=2.31.0

# 유틸리티
tqdm>=4.65.0
python-dotenv>=1.0.0
